+++ NEWS.md
@@ -1,5 +1,7 @@
 # lingglosses 0.0.3
 
+- rename `orthography` to `annotation` in the `gloss_example()` function.
+
 # lingglosses 0.0.2
 
 - change `glosses` dataset to `glosses_df` (to avoid the same name with the `gloss_example()`'s argument)